First pass claim all modal
